Address review comments for microllama and fix starcoder crash
