+++ litgpt/data/microllama.py
@@ -3,10 +3,10 @@ from dataclasses import dataclass
 from pathlib import Path
 from typing import Union
 
-from litgpt.data import LlamaDataModule
+from litgpt.data import TinyLlama
 
 @dataclass
-class MicroLlama(LlamaDataModule):
+class MicroLlama(TinyLlama):
     """The MicroLlama data module is composed of only SlimPajama data."""
 
     def __init__(self, data_path: Union[str, Path] = Path("data/"), seed: int = 42, num_workers: int = 8):
